Fix worker desktopthread error (#155)
* More SetThreadDesktop

* Worer failure 1

* Worker failed 2

+++ lanthing/src/worker/worker_check_decode.cpp
@@ -76,8 +76,8 @@ WorkerCheckDecode::WorkerCheckDecode(uint32_t codecs)
 
 WorkerCheckDecode::~WorkerCheckDecode() = default;
 
-uint32_t WorkerCheckDecode::wait() {
-    return codecs_;
+int WorkerCheckDecode::wait() {
+    return static_cast<int>(codecs_);
 }
 
 } // namespace worker
